requirements.txt
asgiref==3.2.10
attrs==22.1.0
certifi==2021.10.8
chardet==5.1.0
charset-normalizer==2.0.7
django-csvimport==2.16
django-filter==21.1
Django==2.2.16
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
djoser
drf-base64==2.0
Faker==12.0.1
Flake8==4.0.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
isort==5.11.5
mixer==7.1.2
packaging==21.3
Pillow==8.3.1
pip==23.0.1
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.11.0
PyJWT==2.6.0
pyparsing==3.0.9
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
requests==2.28.2
reportlab
setuptools==60.2.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.3.1
toml==0.10.2
typing-extensions==4.4.0
urllib3==1.26.7
wheel==0.37.1
whitenoise==5.2.0
zipp==3.9.0
